Fix implicit conversion warnings

+++ tests/extract_chilitag_dict.cpp
@@ -75,7 +75,7 @@ int main(int argc, char** argv)
     // insert the elements
     cout << "\n\nvector<uint64_t> codes={";
     int idx = 0;
-    for (int i = 0; i < numbers.size(); i++)
+    for (size_t i = 0; i < numbers.size(); i++)
     {
         cout << "0x" << std::hex << numbers[i] << "UL";
         if (i < numbers.size() - 1)
